requirements.txt
# requirements.txt

altgraph==0.17.4
APScheduler==3.11.0
beautifulsoup4==4.13.4
bidict==0.23.1
blinker==1.9.0
bottle==0.13.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
clr_loader==0.2.7.post0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Flask==3.1.0
Flask-APScheduler==1.13.1
Flask-Assets==2.1.0
flask-cors==5.0.1
Flask-Scss==0.5
Flask-SocketIO==5.5.1
fonttools==4.57.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
keyring==25.6.0
kiwisolver==1.4.8
libsass==0.23.0
luadata==1.0.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.7.0
Nuitka==2.6.9
numpy==2.2.4
ordered-set==4.1.0
packaging==24.2
pefile==2023.2.7
pillow==11.2.1
proxy_tools==0.1.0
psutil==7.0.0
pycparser==2.22
Pygments==2.19.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyScss==1.4.0
python-dateutil==2.9.0.post0
python-engineio==4.12.0
python-socketio==5.13.0
pythonnet==3.0.5
pywebview==5.4
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
setuptools==78.1.0
simple-websocket==1.1.0
six==1.17.0
soupsieve==2.6
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
webassets==2.0
Werkzeug==3.1.3
wsproto==1.2.0
zstandard==0.23.0
